# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/kategoris": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Display a listing of all categories",
                "description": "Get a list of all categories",
                "operationId": "indexKategori",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "nama_kategori": {
                                                "type": "string"
                                            },
                                            "tipe": {
                                                "type": "string"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Store a newly created category",
                "description": "Create a new category",
                "operationId": "storeKategori",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori",
                                    "tipe"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Food"
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "Pemasukan",
                                            "Pengeluaran"
                                        ]
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nama_kategori": {
                                            "type": "string"
                                        },
                                        "tipe": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/kategoris/{id}": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Display the specified category",
                "description": "Get category by ID",
                "operationId": "showKategori",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nama_kategori": {
                                            "type": "string"
                                        },
                                        "tipe": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Update the specified category",
                "description": "Update an existing category by ID",
                "operationId": "updateKategori",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori",
                                    "tipe"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Transportation"
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "Pemasukan",
                                            "Pengeluaran"
                                        ]
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nama_kategori": {
                                            "type": "string"
                                        },
                                        "tipe": {
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Remove the specified category",
                "description": "Delete a category by ID",
                "operationId": "destroyKategori",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
...
}